Start doing rearrangement work to make discover/collect more singularly focused
