+++ client/package.json
@@ -17,13 +17,13 @@
   "author": "jsasz <justenrickert@gmail.com>",
   "license": "MIT",
   "dependencies": {
+    "@reduxjs/toolkit": "^1.0.4",
     "lodash.shuffle": "^4.2.0",
     "lodash.throttle": "^4.1.1",
     "node-html-parser": "^1.1.16",
     "prettier": "^1.18.2",
     "puppeteer": "^1.20.0",
     "redux": "^4.0.4",
-    "redux-starter-kit": "^1.0.1",
     "sentiment": "^5.0.2"
   },
   "devDependencies": {
