+++ server/routes/utils.js
@@ -10,6 +10,8 @@ const {
   isTheNationHref,
 } = require('../../shared/predicates')
 
+const { parseSite } = require('../../shared/utils')
+
 const {
   BREITBART,
   CNN,
@@ -21,26 +23,6 @@ const {
   VICE,
   VOX,
 } = require('../../shared/constants')
-
-const constant = x => () => x
-
-const cond = matches => x => {
-  if (!matches.length) return undefined
-  if (matches[0][0](x)) return matches[0][1](x)
-  return cond(matches.slice(1))(x)
-}
-
-const parseSite = cond([
-  [isBreitbartHref, constant(BREITBART)],
-  [isCnnHref, constant(CNN)],
-  [isFoxHref, constant(FOX)],
-  [isNprHref, constant(NPR)],
-  [isNbcHref, constant(NBC)],
-  [isTheNationHref, constant(THE_NATION)],
-  [isTheInterceptHref, constant(THE_INTERCEPT)],
-  [isViceHref, constant(VICE)],
-  [isVoxHref, constant(VOX)],
-])
 
 const omit = (o, keys) =>
   Object.keys(o)
